Move Notify to Mailer::Notify
