+++ app/views/mailer/notify/monthly_report_registration.html.slim
